LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/drawing-rectangles-to-the-canvas/2d.fillRect.nonfinite.worker-expected.txt


PASS fillRect() with Infinity/NaN is ignored 
